+++ src/main/java/faang/school/projectservice/controller/recommendation/ProjectController.java
@@ -8,15 +8,22 @@ import faang.school.projectservice.dto.project.ProjectUpdateDto;
 import faang.school.projectservice.service.ProjectManagementService;
 import faang.school.projectservice.service.ProjectService;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.InputStream;
 import java.util.List;
 
 @RestController
@@ -63,4 +70,19 @@ public class ProjectController {
         return projectService.getPresentationFileKey(projectId);
     }
 
+    @PutMapping("/{projectId}/upload/cover")
+    public void addCoverToProject(@PathVariable @Min(1) long projectId,
+                                  @RequestParam("cover") @NotNull MultipartFile cover) {
+        projectService.addCoverToProject(projectId, cover);
+    }
+
+    @GetMapping("/{projectId}/cover")
+    public InputStream getCover(@PathVariable @Min(1) long projectId) {
+        return projectService.getCover(projectId);
+    }
+
+    @DeleteMapping("/{projectId}/cover")
+    public void deleteAvatar(@PathVariable @Min(1) long projectId) {
+        projectService.deleteCover(projectId);
+    }
 }
